+++ feature/experimental/p4rt/ate_tests/google_discovery_protocol_packetout_test/packetout_test.go
@@ -25,7 +25,7 @@ import (
 )
 
 type PacketIO interface {
-	GetPacketOut(t *testing.T, portID uint32, submitIngress bool) []*p4_v1.PacketOut
+	GetPacketOut(portID uint32, submitIngress bool) []*p4_v1.PacketOut
 }
 
 type testArgs struct {
